CascadeProjects/Azure information/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
fastapi==0.109.0
uvicorn==0.27.0
python-dateutil==2.8.2
jinja2==3.1.3
aiohttp==3.8.5
feedparser==6.0.10
msal==1.26.0
azure-identity==1.15.0
msgraph-sdk==1.0.0
python-dotenv==1.0.0
flask==3.0.0
cryptography==41.0.7
